% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.TSVC.R
\name{confint}
\alias{confint}
\title{Confidence intervals for Varying Coefficient Trees}
\usage{
confint(object, ...)
}
\arguments{
\item{object}{a fitted object of class \code{\link[TSVC]{TSVC}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Confidence intervals for Varying Coefficient Trees
}
